# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for SMPN 1 Dumai website
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guru:
    get:
      operationId: listGuru
      tags: [guru]
      summary: List teachers and staff
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: mata_pelajaran
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            default: nama
        - name: order
          in: query
          schema:
            type: string
            default: asc
      responses:
        "200":
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guru"
    post:
      operationId: createGuru
      tags: [guru]
      summary: Create teacher (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuruBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guru"
  /guru/{id}:
    get:
      operationId: getGuruById
      tags: [guru]
      summary: Get teacher by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guru"
    put:
      operationId: updateGuru
      tags: [guru]
      summary: Update teacher (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuruBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guru"
    delete:
      operationId: deleteGuru
      tags: [guru]
      summary: Delete teacher (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ─── GALERI ──────────────────────────────────────────────────────────────────
...
  /pengumuman:
    get:
      operationId: listPengumuman
      tags: [pengumuman]
      summary: List announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pengumuman"
    post:
      operationId: createPengumuman
      tags: [pengumuman]
      summary: Create announcement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePengumumanBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pengumuman"
  /pengumuman/{id}:
    put:
      operationId: updatePengumuman
      tags: [pengumuman]
      summary: Update announcement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePengumumanBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pengumuman"
    delete:
      operationId: deletePengumuman
      tags: [pengumuman]
      summary: Delete announcement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ─── PRESTASI ────────────────────────────────────────────────────────────────
  /prestasi:
    get:
      operationId: listPrestasi
      tags: [prestasi]
      summary: List achievements
      parameters:
        - name: kategori
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prestasi"
    post:
      operationId: createPrestasi
      tags: [prestasi]
      summary: Create achievement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrestasiBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prestasi"
  /prestasi/{id}:
    put:
      operationId: updatePrestasi
      tags: [prestasi]
      summary: Update achievement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrestasiBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prestasi"
    delete:
      operationId: deletePrestasi
      tags: [prestasi]
      summary: Delete achievement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ─── EKSKUL ──────────────────────────────────────────────────────────────────
...
  /slider:
    get:
      operationId: listSlider
      tags: [slider]
      summary: List hero slider images
      responses:
        "200":
          description: Slider images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Slider"
    post:
      operationId: createSlider
      tags: [slider]
      summary: Add slider image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSliderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Slider"
  /slider/{id}:
    delete:
      operationId: deleteSlider
      tags: [slider]
      summary: Delete slider image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ─── SPMB ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # ── Berita ─────────────────────────────────────────────────────────────────
...
    Guru:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        nip:
          type: string
          nullable: true
        mata_pelajaran:
          type: string
        jabatan:
          type: string
        foto:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
      required:
        - id
        - nama
        - mata_pelajaran
        - jabatan
        - is_active
        - created_at
    CreateGuruBody:
      type: object
      properties:
        nama:
          type: string
        nip:
          type: string
          nullable: true
        mata_pelajaran:
          type: string
        jabatan:
          type: string
        foto:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        is_active:
          type: boolean
      required:
        - nama
        - mata_pelajaran
        - jabatan
        - is_active

    # ── Galeri ─────────────────────────────────────────────────────────────────
...
    Pengumuman:
      type: object
      properties:
        id:
          type: integer
        judul:
          type: string
        konten:
          type: string
        prioritas:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
      required:
        - id
        - judul
        - konten
        - prioritas
        - is_active
        - created_at
    CreatePengumumanBody:
      type: object
      properties:
        judul:
          type: string
        konten:
          type: string
        prioritas:
          type: integer
        is_active:
          type: boolean
      required:
        - judul
        - konten
        - prioritas
        - is_active

    # ── Prestasi ───────────────────────────────────────────────────────────────
    Prestasi:
      type: object
      properties:
        id:
          type: integer
        judul:
          type: string
        deskripsi:
          type: string
          nullable: true
        tanggal:
          type: string
        tingkat:
          type: string
        kategori:
          type: string
        created_at:
          type: string
      required:
        - id
        - judul
        - tanggal
        - tingkat
        - kategori
        - created_at
    CreatePrestasiBody:
      type: object
      properties:
        judul:
          type: string
        deskripsi:
          type: string
          nullable: true
        tanggal:
          type: string
        tingkat:
          type: string
        kategori:
          type: string
      required:
        - judul
        - tanggal
        - tingkat
        - kategori

    # ── Ekskul ─────────────────────────────────────────────────────────────────
...
    Slider:
      type: object
      properties:
        id:
          type: integer
        judul:
          type: string
        subjudul:
          type: string
          nullable: true
        gambar:
          type: string
        urutan:
          type: integer
        is_active:
          type: boolean
      required:
        - id
        - judul
        - gambar
        - urutan
        - is_active
    CreateSliderBody:
      type: object
      properties:
        judul:
          type: string
        subjudul:
          type: string
          nullable: true
        gambar:
          type: string
        urutan:
          type: integer
        is_active:
          type: boolean
      required:
        - judul
        - gambar
        - urutan
        - is_active

    # ── SPMB ───────────────────────────────────────────────────────────────────
...