# lib/api-zod/src/generated/types/createGuruBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SMPN 1 Dumai website
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGuruBody {
  nama: string;
  nip?: string | null;
  mata_pelajaran: string;
  jabatan: string;
  foto?: string | null;
  bio?: string | null;
  is_active: boolean;
}