# lib/api-zod/src/generated/types/guru.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SMPN 1 Dumai website
 * OpenAPI spec version: 0.1.0
 */

export interface Guru {
  id: number;
  nama: string;
  nip?: string | null;
  mata_pelajaran: string;
  jabatan: string;
  foto?: string | null;
  bio?: string | null;
  is_active: boolean;
  created_at: string;
}